configs: tcu_fs: Use 128-bit bus widths for KecAcc tiles
Configure more realistic bus withs for tiles with the KecAcc hash
accelerator by setting all bus widths to 128-bit. This matches what the
accelerator makes use of by default and is a more conservative choice.

To make this work properly, the "cache line size" of the CPU should also
be adjusted so it doesn't make overly large accesses for instruction
fetches or unaligned data.

NOTE: Changing the cache line size does not seem to work properly for
the O3CPU (leads to gem5 crashes), so it is limited to the MinorCPU for
now.

Change-Id: Ie07b9f1ebe5a8d1c975ce1c0aec6b6256c202a4f