configs: tcu_fs: Split HashMux-specific part of createKecAccTile
The RoT tile does not make use of double buffering at the moment, so the
FIXME comment is not relevant for it. The accelerator can just be
connected to the XBar like everything else.

Also, multiple Scratchpads can now be added to a tile, so it would be
possible to revisit this and simulate it properly with separate SRAMs.
Add a new TODO line for that.

Change-Id: Idf0c2fcb0f050318c9e639b0fb24b78d8eedf8df