configs: tcu_fs: Adjust RoT tile for gem5 v23.1.0.0
gem5 v23.1.0.0 provides configuration options to limit the functionality
of RISC-V cores. Since the RoT tile is designed to have a very simple
CPU core, it is more realistic to disable unneeded functionality. This
works fine for the RISC-V Vector (RVV) extension. The set of privilege
levels cannot be limited yet at the moment, because TileMux is necessary
for running the RoT service and it does not work without supervisor
mode. The RoT layers work fine with only M-mode.

Change-Id: I794303a9c1d11f91216989096ffa4b2ff23ed1e2