sim: m3_loader: Configure kernel virteps count for RoT
With recent changes to the TCU the RoT needs to set up the memory region
that will be used for the virtual endpoints of the kernel. This should
be configurable and included in the signed M3 configuration report.

Add the option to the ROSA configuration structure. For now just
configure the kernel tile with the same amount of EPs as used for the
RoT tile (could also make this separately configurable).

The configuration only includes the number of *pages* for the EP region
because the current implementation in ROSA can only allocate full pages
in memory anyway and this saves both space and validation effort.

Change-Id: Ib6f2e6be6750c502b3b09c34f1f981504050cfbd